Implement setting a `ygoutput_xpub` for a wallet.
Added two new wallet tool commands: `setygoutputxpub` and `showygoutputxpub`.

This `ygoutput_xpub` is an xpub of an external HD wallet (like a cold storage wallet). By storing the `xpub` of that external wallet, we can monitor the wallet's addresses, and send our yield generator coin-join output directly to an unused address of the cold storage wallet.